Philosopher: Philosopher Routine

+++ philo/src/utils/time_management_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   time_management_utils.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gyasminalves <gyasminalves@student.42.f    +#+  +:+       +#+        */
+/*   By: galves-a <galves-a@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/02 19:46:02 by galves-a          #+#    #+#             */
-/*   Updated: 2025/06/03 18:07:26 by gyasminalve      ###   ########.fr       */
+/*   Updated: 2025/06/12 19:15:30 by galves-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 long long   get_time_in_ms(void)
 {
     struct timeval timestamp;
-    long    time_in_ms;
+    long long    time_in_ms;
     
     gettimeofday(&timestamp, NULL);
     time_in_ms = (timestamp.tv_sec * 1000) + (timestamp.tv_usec / 1000);
